BAEL-4198 - Fix Selenium Testing module (#9611)
* BAEL-4198 - Fix Selenium Live Tests

* BAEL-4198
