+++ testing-modules/selenium-junit-testng/src/main/java/com/baeldung/selenium/SeleniumExample.java
@@ -1,14 +1,12 @@
-package main.java.com.baeldung.selenium;
-
-import main.java.com.baeldung.selenium.config.SeleniumConfig;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.interactions.Actions;
+package com.baeldung.selenium;
 
 import java.util.List;
-import java.util.concurrent.TimeUnit;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+
+import com.baeldung.selenium.config.SeleniumConfig;
 
 public class SeleniumExample {
 
@@ -17,15 +15,18 @@ public class SeleniumExample {
 
     public SeleniumExample() {
         config = new SeleniumConfig();
-        config.getDriver().get(url);
+        config.getDriver()
+            .get(url);
     }
 
     public void closeWindow() {
-        this.config.getDriver().close();
+        this.config.getDriver()
+            .close();
     }
 
     public String getTitle() {
-        return this.config.getDriver().getTitle();
+        return this.config.getDriver()
+            .getTitle();
     }
 
     public void getAboutBaeldungPage() {
@@ -35,29 +36,35 @@ public class SeleniumExample {
     }
 
     private void closeOverlay() {
-        List<WebElement> webElementList = this.config.getDriver().findElements(By.tagName("a"));
+        List<WebElement> webElementList = this.config.getDriver()
+            .findElements(By.tagName("a"));
         if (webElementList != null) {
             webElementList.stream()
-              .filter(webElement -> "Close".equalsIgnoreCase(webElement.getAttribute("title")))
-              .filter(WebElement::isDisplayed)
-              .findAny()
-              .ifPresent(WebElement::click);
+                .filter(webElement -> "Close".equalsIgnoreCase(webElement.getAttribute("title")))
+                .filter(WebElement::isDisplayed)
+                .findAny()
+                .ifPresent(WebElement::click);
         }
     }
 
     private void clickAboutLink() {
-        this.config.getDriver().findElement(By.partialLinkText("About")).click();
+        Actions actions = new Actions(config.getDriver());
+        WebElement aboutElement = this.config.getDriver()
+            .findElement(By.id("menu-item-6138"));
+        
+        actions.moveToElement(aboutElement).perform();
     }
 
     private void clickAboutUsLink() {
-        Actions builder = new Actions(config.getDriver());
-        WebElement element = this.config.getDriver().findElement(By.partialLinkText("About Baeldung."));
-        builder.moveToElement(element).build().perform();
+        WebElement element = this.config.getDriver()
+            .findElement(By.partialLinkText("About Baeldung."));
+        element.click();
     }
 
     public boolean isAuthorInformationAvailable() {
         return this.config.getDriver()
-                .findElement(By.cssSelector("article > .row > div"))
-                .isDisplayed();
+            .getPageSource()
+            .contains("Hey ! I'm Eugen");
     }
+
 }
