+++ testing-modules/selenium-junit-testng/src/main/java/com/baeldung/selenium/config/SeleniumConfig.java
@@ -1,16 +1,13 @@
-package main.java.com.baeldung.selenium.config;
+package com.baeldung.selenium.config;
+
+import java.io.File;
+import java.util.concurrent.TimeUnit;
 
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
-
-import java.io.File;
-import java.util.concurrent.TimeUnit;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class SeleniumConfig {
 
@@ -19,15 +16,17 @@ public class SeleniumConfig {
     public SeleniumConfig() {
         Capabilities capabilities = DesiredCapabilities.firefox();
         driver = new FirefoxDriver(capabilities);
-        driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
+        driver.manage()
+            .timeouts()
+            .implicitlyWait(5, TimeUnit.SECONDS);
     }
 
     static {
         System.setProperty("webdriver.gecko.driver", findFile("geckodriver.mac"));
     }
 
-    static private String findFile(String filename) {
-        String paths[] = {"", "bin/", "target/classes"}; // if you have chromedriver somewhere else on the path, then put it here.
+    private static String findFile(String filename) {
+        String[] paths = { "", "bin/", "target/classes" }; // if you have chromedriver somewhere else on the path, then put it here.
         for (String path : paths) {
             if (new File(path + filename).exists())
                 return path + filename;
@@ -40,7 +39,8 @@ public class SeleniumConfig {
     }
 
     public void navigateTo(String url) {
-        driver.navigate().to(url);
+        driver.navigate()
+            .to(url);
     }
 
     public void clickElement(WebElement element) {
