+++ testing-modules/selenium-junit-testng/src/main/java/com/baeldung/selenium/models/BaeldungAbout.java
@@ -1,7 +1,7 @@
-package main.java.com.baeldung.selenium.models;
+package com.baeldung.selenium.models;
 
-import main.java.com.baeldung.selenium.config.SeleniumConfig;
-import main.java.com.baeldung.selenium.pages.BaeldungAboutPage;
+import com.baeldung.selenium.config.SeleniumConfig;
+import com.baeldung.selenium.pages.BaeldungAboutPage;
 import org.openqa.selenium.support.PageFactory;
 
 public class BaeldungAbout {
