+++ testing-modules/selenium-junit-testng/src/main/java/com/baeldung/selenium/pages/BaeldungAboutPage.java
@@ -1,4 +1,4 @@
-package main.java.com.baeldung.selenium.pages;
+package com.baeldung.selenium.pages;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
