+++ testing-modules/selenium-junit-testng/src/test/java/com/baeldung/selenium/junit/SeleniumWithJUnitLiveTest.java
@@ -1,16 +1,21 @@
-package test.java.com.baeldung.selenium.junit;
+package com.baeldung.selenium.junit;
 
-import main.java.com.baeldung.selenium.SeleniumExample;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
+import java.io.IOException;
+
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import static org.testng.Assert.*;
+import com.baeldung.selenium.SeleniumExample;
 
 public class SeleniumWithJUnitLiveTest {
 
     private static SeleniumExample seleniumExample;
-    private String expectedTitle = "Baeldung | Java, Spring and Web Development tutorials";
+    private String expectedTitle = "About Baeldung | Baeldung";
 
     @BeforeClass
     public static void setUp() {
@@ -18,17 +23,17 @@ public class SeleniumWithJUnitLiveTest {
     }
 
     @AfterClass
-    public static void tearDown() {
+    public static void tearDown() throws IOException {
         seleniumExample.closeWindow();
     }
 
     @Test
     public void whenAboutBaeldungIsLoaded_thenAboutEugenIsMentionedOnPage() {
-            seleniumExample.getAboutBaeldungPage();
-            String actualTitle = seleniumExample.getTitle();
-            assertNotNull(actualTitle);
-            assertEquals(expectedTitle, actualTitle);
-            assertTrue(seleniumExample.isAuthorInformationAvailable());
+        seleniumExample.getAboutBaeldungPage();
+        String actualTitle = seleniumExample.getTitle();
+        assertNotNull(actualTitle);
+        assertEquals(expectedTitle, actualTitle);
+        assertTrue(seleniumExample.isAuthorInformationAvailable());
     }
 
 }
